Remove the Event class (only keep InGameEvent).
The Event class did not really have the correct structure to represent
lobby events, so the inheritance was not justified. This class was
indeed based on the assumption that communication would happen over UDP
(not the case when in lobby), that every event would need a timestamp
and an ID (useless for the lobby), etc.

This commit also removes all existing lobby events in order to have a
clean state to properly implement them.